+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+# [1.3.0-beta.5](https://github.com/TouK/federated-types/compare/v1.3.0-beta.4...v1.3.0-beta.5) (2026-05-28)
+
+
+### Bug Fixes
+
+* trigger beta release to verify OIDC after secret cleanup ([41fcc4e](https://github.com/TouK/federated-types/commit/41fcc4e93128f05e37968913b187c8303bddc07f))
+
+# [1.3.0-beta.4](https://github.com/TouK/federated-types/compare/v1.3.0-beta.3...v1.3.0-beta.4) (2026-05-28)
+
+
+### Bug Fixes
+
+* use canonical TouK case in repository URL for npm provenance ([908e120](https://github.com/TouK/federated-types/commit/908e120becf3ace91d45faf09da4324ac533eb29))
+
+# [1.3.0-beta.3](https://github.com/touk/federated-types/compare/v1.3.0-beta.2...v1.3.0-beta.3) (2026-05-28)
+
+
+### Bug Fixes
+
+* trigger release after orphaned 1.3.0-beta.3 tag ([61ddeb3](https://github.com/touk/federated-types/commit/61ddeb34bc0687a36a519ab3524b170a793dfcb8))
+* use correct npm config key for min-release-age ([a50c721](https://github.com/touk/federated-types/commit/a50c7212394ef9af1f50ebbd4a0753d012dbae0c))
+
+
+### Performance Improvements
+
+* cache TS build info for incremental type generation ([e8ab9e6](https://github.com/touk/federated-types/commit/e8ab9e633e450fd7d0d2705466b712f08d9d1cc6))
+
+# [1.3.0-beta.2](https://github.com/touk/federated-types/compare/v1.3.0-beta.1...v1.3.0-beta.2) (2026-03-10)
+
+
+### Features
+
+* add colored output with chalk ([1b33197](https://github.com/touk/federated-types/commit/1b3319781edf5b2f900026dbcf718d1cbcaae515))
+
+# [1.3.0-beta.1](https://github.com/touk/federated-types/compare/v1.2.2-beta.1...v1.3.0-beta.1) (2026-03-10)
+
+
+### Bug Fixes
+
+* extension in exposed path ([e707c3e](https://github.com/touk/federated-types/commit/e707c3e8fb6aa53f0f9a961cf5b425b09f2626f6))
+
+
+### Features
+
+* add inline config via --name and --exposes CLI arguments ([34e4752](https://github.com/touk/federated-types/commit/34e4752e2c5b2d35e788b82faf8d463981b29368))
+
+## [1.2.2-beta.1](https://github.com/touk/federated-types/compare/v1.2.1...v1.2.2-beta.1) (2026-02-19)
+
+
+### Bug Fixes
+
+* ci build ([a3f874f](https://github.com/touk/federated-types/commit/a3f874fb81da3f056b8cf20e714740cc3a03afa5))
+* update major deps ([b7f5168](https://github.com/touk/federated-types/commit/b7f516815c8da61760da3746ce47455c5c968c72))
+
 ## [1.2.1](https://github.com/touk/federated-types/compare/v1.2.0...v1.2.1) (2025-06-28)
 
 
@@ -14,68 +68,58 @@
 
 # [1.2.0](https://github.com/touk/federated-types/compare/v1.1.0...v1.2.0) (2022-10-31)
 
-
 ### Bug Fixes
 
-* npm audit fix ([e95a125](https://github.com/touk/federated-types/commit/e95a125541584a95038d209f285ef28ea60a38d2))
-
+-   npm audit fix ([e95a125](https://github.com/touk/federated-types/commit/e95a125541584a95038d209f285ef28ea60a38d2))
 
 ### Features
 
-* "--saveToNodeModules" param ([8ca1214](https://github.com/touk/federated-types/commit/8ca1214759cc2659e546c5e99d3121374947be4b))
+-   "--saveToNodeModules" param ([8ca1214](https://github.com/touk/federated-types/commit/8ca1214759cc2659e546c5e99d3121374947be4b))
 
 # [1.2.0-beta.1](https://github.com/touk/federated-types/compare/v1.1.0...v1.2.0-beta.1) (2022-10-31)
 
-
 ### Bug Fixes
 
-* npm audit fix ([e95a125](https://github.com/touk/federated-types/commit/e95a125541584a95038d209f285ef28ea60a38d2))
-
+-   npm audit fix ([e95a125](https://github.com/touk/federated-types/commit/e95a125541584a95038d209f285ef28ea60a38d2))
 
 ### Features
 
-* "--saveToNodeModules" param ([8ca1214](https://github.com/touk/federated-types/commit/8ca1214759cc2659e546c5e99d3121374947be4b))
+-   "--saveToNodeModules" param ([8ca1214](https://github.com/touk/federated-types/commit/8ca1214759cc2659e546c5e99d3121374947be4b))
 
 # [1.1.0](https://github.com/touk/federated-types/compare/v1.0.0...v1.1.0) (2021-08-02)
 
-
 ### Features
 
-* use custom config path with --config arg ([e85dea5](https://github.com/touk/federated-types/commit/e85dea5994b8f8a0ff050fe0c62c8751fdd20bcd))
+-   use custom config path with --config arg ([e85dea5](https://github.com/touk/federated-types/commit/e85dea5994b8f8a0ff050fe0c62c8751fdd20bcd))
 
 # [1.1.0-beta.1](https://github.com/touk/federated-types/compare/v1.0.0...v1.1.0-beta.1) (2021-08-02)
 
-
 ### Features
 
-* use custom config path with --config arg ([e85dea5](https://github.com/touk/federated-types/commit/e85dea5994b8f8a0ff050fe0c62c8751fdd20bcd))
+-   use custom config path with --config arg ([e85dea5](https://github.com/touk/federated-types/commit/e85dea5994b8f8a0ff050fe0c62c8751fdd20bcd))
 
 # 1.0.0 (2021-07-21)
 
-
 ### Features
 
-* aliases support, cleanup ([064eb35](https://github.com/touk/federated-types/commit/064eb35d608752ecb5b2bbd0c27bdc149c1c31b7))
-* inital release ([b4b8e37](https://github.com/touk/federated-types/commit/b4b8e371decf0842774834682c7391496720c1d6))
-* report ts diagnostics ([8cda361](https://github.com/touk/federated-types/commit/8cda3618d052c21876815940d55405521fa7ad68))
+-   aliases support, cleanup ([064eb35](https://github.com/touk/federated-types/commit/064eb35d608752ecb5b2bbd0c27bdc149c1c31b7))
+-   inital release ([b4b8e37](https://github.com/touk/federated-types/commit/b4b8e371decf0842774834682c7391496720c1d6))
+-   report ts diagnostics ([8cda361](https://github.com/touk/federated-types/commit/8cda3618d052c21876815940d55405521fa7ad68))
 
 # [1.0.0-beta.3](https://github.com/touk/federated-types/compare/v1.0.0-beta.2...v1.0.0-beta.3) (2021-07-21)
 
-
 ### Features
 
-* aliases support, cleanup ([064eb35](https://github.com/touk/federated-types/commit/064eb35d608752ecb5b2bbd0c27bdc149c1c31b7))
+-   aliases support, cleanup ([064eb35](https://github.com/touk/federated-types/commit/064eb35d608752ecb5b2bbd0c27bdc149c1c31b7))
 
 # [1.0.0-beta.2](https://github.com/touk/federated-types/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2021-07-21)
 
-
 ### Features
 
-* report ts diagnostics ([8cda361](https://github.com/touk/federated-types/commit/8cda3618d052c21876815940d55405521fa7ad68))
+-   report ts diagnostics ([8cda361](https://github.com/touk/federated-types/commit/8cda3618d052c21876815940d55405521fa7ad68))
 
 # 1.0.0-beta.1 (2021-07-21)
 
-
 ### Features
 
-* inital release ([b4b8e37](https://github.com/touk/federated-types/commit/b4b8e371decf0842774834682c7391496720c1d6))
+-   inital release ([b4b8e37](https://github.com/touk/federated-types/commit/b4b8e371decf0842774834682c7391496720c1d6))